Administracion de usuarios y sesiones, actualizacion de control de usuarios en sistema

+++ roles/delete.php
@@ -4,30 +4,40 @@ ini_set('display_errors', '1');
 
 require('../class/config.php');
 require('../class/rutas.php');
+require('../class/session.php');
 require('../class/rolModel.php');
 
-if (isset($_POST['confirm']) && $_POST['confirm'] == 1) {
-    $id = (int) $_POST['rol'];
+$session = new Session;
 
-    //print_r($id);exit;
+if (isset($_SESSION['autenticado']) && $_SESSION['usuario_rol'] == 'Administrador') {
+    # code...
 
-    #verificar que hay un rol con el id recibido
-    $roles = new RolModel;
-    $rol = $roles->getRolId($id);
 
-    if ($rol) {
-        #eliminar el rol
-        $row = $roles->deleteRol($id);
+    if (isset($_POST['confirm']) && $_POST['confirm'] == 1) {
+        $id = (int) $_POST['rol'];
 
-        if ($row) {
-            $_SESSION['success'] = 'El rol se ha eliminado correctamente';
+        //print_r($id);exit;
+
+        #verificar que hay un rol con el id recibido
+        $roles = new RolModel;
+        $rol = $roles->getRolId($id);
+
+        if ($rol) {
+            #eliminar el rol
+            $row = $roles->deleteRol($id);
+
+            if ($row) {
+                $_SESSION['success'] = 'El rol se ha eliminado correctamente';
+                header('Location: ' . ROLES);
+            }else {
+                $_SESSION['danger'] = 'El rol no se ha podido eliminar... intente mas tarde';
+                header('Location: ' . SHOW_ROL . $id);
+            }
+        }else{
+            $_SESSION['danger'] = 'El rol no está registrado';
             header('Location: ' . ROLES);
-        }else {
-            $_SESSION['danger'] = 'El rol no se ha podido eliminar... intente mas tarde';
-            header('Location: ' . SHOW_ROL . $id);
         }
-    }else{
-        $_SESSION['danger'] = 'El rol no está registrado';
-        header('Location: ' . ROLES);
     }
+}else {
+    header('Location: ' . LOGIN);
 }
